Move condaLocator.ts
