+++ src/client/pythonEnvironments/base/locators/lowLevel/condaLocator.ts
@@ -1,17 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import '../../../../common/extensions';
-import { PythonEnvKind } from '../../../base/info';
-import { BasicEnvInfo, IPythonEnvsIterator, Locator } from '../../../base/locator';
+import { PythonEnvKind } from '../../info';
+import { BasicEnvInfo, IPythonEnvsIterator, Locator } from '../../locator';
 import { getInterpreterPathFromDir } from '../../../common/commonUtils';
 import { Conda } from '../../../common/environmentManagers/conda';
 import { traceError, traceVerbose } from '../../../../common/logger';
 
 export class CondaEnvironmentLocator extends Locator<BasicEnvInfo> {
-    public constructor() {
-        super();
-    }
-
     // eslint-disable-next-line class-methods-use-this
     public async *iterEnvs(): IPythonEnvsIterator<BasicEnvInfo> {
         const conda = await Conda.getConda();
